requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.6
django-crispy-forms==2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
